nix: Make carla package independent of my nix-conf repo
Compared to the previous version of the package, this package does not
contain python stuff as the Carla Python module is now available
independently in nix/carla/carla_py.nix. Thanks to this, the Carla
package can be built quite fast. Previously, Python had to be built
from source, which took long time.